Fix task 9
The last version of task 9 showed too many invalid activities, as there
was no check for an occurence of a new activity. That means, if the last activity's timestamp was in 2008, and the first timestamp in the new activity was in 2015, it would be marked as invalid since there are 7 years inbetween. This issue is solved with this commit.